Flex patch so both xenial and arch are happy. (htj): Note that this goes in line with the other patching, making int -> size_t. That is however going against the current trend in flex, which goes size_t -> int.
That a battle we cannot really win, see:

https://github.com/westes/flex/commit/4a8eccf830e36f6611dc9cc6debd815ff32dcaf5 [github.com]
https://github.com/westes/flex/commit/a51bd6c3016176b83d30fc44818d4ba3b8c946ba [github.com]

Now, the commit message of the first commit there makes me worried.
If this is really a buffer size, it should be some size_t and not int...
These are in flex v2.6.1.

Have to come up with a different patch strategy.